faq/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
requests==2.26.0
pymysql==1.0.2
django-cors-headers==3.7.0
django-filter==2.4.0